backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
passlib==1.7.4
# passlib[bcrypt]==1.7.4
httpx
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic-settings
pydantic[email]
pandas